+++ .eslintrc.js
@@ -1,4 +1,0 @@
-module.exports = {
-    extends: 'loris/es6',
-    root: true
-};