+++ source/js/functions/debounce.js
@@ -2,7 +2,9 @@
 // be triggered. The function will be called after it stops being called for
 // N milliseconds. If `immediate` is passed, trigger the function on the
 // leading edge, instead of the trailing.
+/* eslint-disable no-unused-vars */
 function debounce(func, wait, immediate) {
+    'use strict';
     var timeout;
     return function () {
         var context = this;