+++ source/js/functions/miss-click.js
@@ -1,5 +1,7 @@
 // Miss click
+/* eslint-disable no-unused-vars */
 jQuery(function ($) {
+    'use strict';
     $(document).mouseup(function (e) { // событие клика по веб-документу
         var div = $('#yourid'); // тут указываем ID элемента
         if (!div.is(e.target) && // если клик был не по нашему блоку