+++ source/js/functions/responsive-iframe.js
@@ -1,9 +1,12 @@
 // Responsive iframe video
 // nowrap - url parametr. For example '//embed.smileexpo'
-function responsiveIframe(contentContainer, nowrap) {
-    var videoWrapper = '<div class="embed-responsive embed-responsive-16by9"></div>';
-    var nowrapIframe = nowrap;
 
-    contentContainer.find('iframe').not("[src ^= nowrap]").wrap(videoWrapper);
+/* eslint-disable no-unused-vars */
+function responsiveIframe(contentContainer, videoWrapper, nowrap) {
+    'use strict';
+
+    videoWrapper = videoWrapper || '<div class="embed-responsive embed-responsive-16by9"></div>';
+
+    contentContainer.find('iframe').not('[src ^=' + nowrap + ']').wrap(videoWrapper);
 }
 // END Responsive iframe video